Move declaration of struct.
