+++ src/fields/assembly_observe.hh
@@ -24,29 +24,8 @@
 #include "coupling/assembly_base.hh"
 #include "fem/dofhandler.hh"
 #include "fem/element_cache_map.hh"
+#include "fem/integral_data.hh"
 #include "io/observe.hh"
-
-
-/**
- * Helper structure holds patch data of observe output
- *
- * Data is specified by cell and subset index in EvalPoint object
- */
-struct ObserveIntegralData {
-	/// Default constructor
-	ObserveIntegralData() {}
-
-    /// Constructor with data mebers initialization
-	ObserveIntegralData(DHCellAccessor dhcell, unsigned int subset_idx)
-    : cell(dhcell), subset_index(subset_idx) {}
-
-    /// Copy constructor
-	ObserveIntegralData(const ObserveIntegralData &other)
-    : cell(other.cell), subset_index(other.subset_index) {}
-
-    DHCellAccessor cell;          ///< Specified cell (element)
-    unsigned int subset_index;    ///< Index (order) of subset in EvalPoints object
-};
 
 
 /**
